Rename scalar/basic_scipy to scalar/math
