+++ aesara/scalar/math.py
