+++ tests/tensor/test_math_opt.py
@@ -2670,7 +2670,7 @@ class TestLocalSwitchSink:
 
 
 @pytest.mark.skipif(
-    config.cxx == "" and not aes.basic_scipy.imported_scipy_special,
+    config.cxx == "" and not aes.math.imported_scipy_special,
     reason="erf need a c++ compiler or scipy",
 )
 class TestLocalErf:
@@ -2762,7 +2762,7 @@ class TestLocalErf:
 
 
 @pytest.mark.skipif(
-    config.cxx == "" and not aes.basic_scipy.imported_scipy_special,
+    config.cxx == "" and not aes.math.imported_scipy_special,
     reason="erf need a c++ compiler or scipy",
 )
 class TestLocalErfc:
